requirements.txt
datasets>=2.19.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
matplotlib>=3.9.0
pyyaml>=6.0.1
tqdm>=4.66.0
python-dotenv>=1.0.1
